Updated
RV_distributions are now stored in a dictionary. This contains the instances for all distributions for a certain class.

Very convenient way now to loop through all the distributions, technically all scipy distributions can be added.

Added the tabulated representation of how good a fit is. Not entirely sure if that is done correctly